+++ CHANGELOG.md
@@ -9,16 +9,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+*
+
+### Changed
+
+*
+
+### Fixed
+
+*
+
+## [0.7.0] - 2026-05-13
+
+### Added
+
 * New `downgrade` command that runs the `rollback` of the last applied migration, raising an error if `rollback` is not implemented
 * New `dry_downgrade` command that prints the last applied migration without rolling it back
 
 ### Changed
 
 * Improved `squash` command description format to use multiline strings with newlines for better readability
-
-### Fixed
-
-*
 
 ## [0.6.0] - 2026-01-07